+++ src/test/kotlin/com/github/victorrentea/livecoding/InspectionParameterizedTestBase.kt
@@ -2,41 +2,46 @@ package com.github.victorrentea.livecoding
 
 import com.intellij.codeInsight.daemon.impl.HighlightInfo
 import com.intellij.codeInspection.LocalInspectionTool
-import com.intellij.testFramework.PsiTestUtil
+import com.intellij.openapi.project.rootManager
 import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase
+import com.intellij.testFramework.fixtures.MavenDependencyUtil
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import java.io.File
 
 @RunWith(Parameterized::class)
-abstract class InspectionParameterizedTestBase(val fileName: String) : LightJavaCodeInsightFixtureTestCase() {
+abstract class InspectionParameterizedTestBase(private val fileName: String) : LightJavaCodeInsightFixtureTestCase() {
     override fun getProjectDescriptor() = JAVA_8
 
     override fun getTestDataPath() = JAVA_SRC_FOLDER
 
-    protected fun HighlightInfo.getLineNumber() = getLineNumber(myFixture.file, actualStartOffset)
+    private fun HighlightInfo.getLineNumber() = getLineNumber(myFixture.file, actualStartOffset)
 
     protected abstract fun inspectionName(): String
     protected abstract fun fixName(): String
 
     protected abstract fun inspectionClass(): Class<out LocalInspectionTool>
 
+
     @Test
     fun test() {
-        myFixture.enableInspections(inspectionClass())
-//        PsiTestUtil.addLibrary(myFixture.module,
-//            """C:\Users\victo\.m2\repository\org\projectlombok\lombok\1.18.22\lombok-1.18.22.jar"""
-//        )
 
-        myFixture.configureByFiles(fileName);
+        MavenDependencyUtil.addFromMaven(module.rootManager.modifiableModel, "org.projectlombok:lombok:1.18.22")
+
+        myFixture.enableInspections(inspectionClass())
+
+        myFixture.configureByFiles(fileName)
 
         val expectedHighlightedLines = File(JAVA_SRC_FOLDER, fileName).readLines()
             .mapIndexedNotNull { lineNumber, line -> if (line.contains("//")) lineNumber + 1 else null }
-        println("EXPECTED HIGHLIGHTED LINES: " + expectedHighlightedLines)
+        println("EXPECTED HIGHLIGHTED LINES: $expectedHighlightedLines")
 
         val highlights = myFixture.doHighlighting()
-            /*.also { println("Highlights: " + it) }.*/
+            .also { allHighlights ->
+                println("-- all highlights --")
+                allHighlights.forEach { println("${it.getLineNumber()}: ${it.description}") }
+            }
             .filter { it.description == inspectionName() }
 
         val actualHighlightedLines = highlights.map { it.getLineNumber() }
@@ -56,7 +61,7 @@ abstract class InspectionParameterizedTestBase(val fileName: String) : LightJava
             if (!expectedFile.isFile) {
                 throw IllegalArgumentException("Expected file not found at ${expectedFile.absolutePath}")
             }
-            myFixture.checkResultByFile(getExpectedFile(fileName));
+            myFixture.checkResultByFile(getExpectedFile(fileName))
         }
     }
 }