+++ src/test/kotlin/com/github/victorrentea/livecoding/LombokAddRacTest.kt
@@ -1,8 +1,9 @@
 package com.github.victorrentea.livecoding
 
-import com.github.victorrentea.livecoding.lombok.AddRequiredArgsConstructorFix
 import com.github.victorrentea.livecoding.lombok.AddRequiredArgsConstructorInspection
 import com.intellij.codeInspection.LocalInspectionTool
+import com.intellij.openapi.project.rootManager
+import com.intellij.testFramework.fixtures.MavenDependencyUtil
 import org.junit.Ignore
 import org.junit.runners.Parameterized.Parameters
 
@@ -11,13 +12,14 @@ class LombokAddRacTest(fileName: String) : InspectionParameterizedTestBase(fileN
     companion object {
         @JvmStatic
         @Parameters(name = "{0}")
-        fun parameters(): List<String> = getInputFilePaths("lombok.addrac")
+        fun parameters(): List<String> = getInputFilePaths("lombok/addrac")
     }
 
     override fun inspectionName() = AddRequiredArgsConstructorInspection.INSPECTION_NAME
-    override fun fixName() = AddRequiredArgsConstructorFix.FIX_NAME
+    override fun fixName() = AddRequiredArgsConstructorInspection.FIX_NAME
 
     override fun inspectionClass(): Class<out LocalInspectionTool> = AddRequiredArgsConstructorInspection::class.java
+
 
 
 }