+++ src/test/live-coding-playground/src/main/java/lombok/addrac/AddRAC2.java
@@ -4,8 +4,4 @@ public class AddRAC2 {
     private final int x; //
     private final int y; //
 
-    public AddRAC2(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
 }