+++ src/test/live-coding-playground/src/test/java/lombok/addrac/AddRAC.java
@@ -1,5 +1,8 @@
 package lombok.addrac;
 
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
 public class AddRAC {
-    private final int x;
+    private final int x; //
 }